avoided using suppressed field for opt-in feature
